Cleanups

+++ caladesi-neo4j-graph/src/main/scala/net/caladesiframework/neo4j/graph/repository/Neo4jGraphRepository.scala
@@ -239,15 +239,20 @@ abstract class Neo4jGraphRepository[EntityType <: Neo4jGraphEntity]
    * @return
    */
   def delete(entity: EntityType) = transactional[Boolean]( implicit ds => {
-    val node = entity.getUnderlyingNode
-    val relations = node.getRelationships(Direction.OUTGOING)
+    def node = entity.getUnderlyingNode
 
-    // Remove all outgoing relations
-    while (relations.iterator().hasNext) {
-      relations.iterator().next().delete()
+    // DO NOT remove incoming relations, this will break data integrity (user has to check before deletion)
+    def relationsIncoming = node.getRelationships(Direction.INCOMING)
+    if (relationsIncoming.iterator().hasNext) {
+      throw new Exception("Can not delete entity, still has incoming connections")
     }
 
-    // Please check all incoming relations in repository
+    // Remove all outgoing relations
+    val relationsOutgoing = node.getRelationships(Direction.OUTGOING)
+    while (relationsOutgoing.iterator().hasNext) {
+      relationsOutgoing.iterator().next().delete()
+    }
+
     removeFromIndex(entity)
     entity.getUnderlyingNode.delete()
     true
@@ -301,20 +306,15 @@ abstract class Neo4jGraphRepository[EntityType <: Neo4jGraphEntity]
   protected def setNodeFields(node: Node, entity: EntityType)(implicit ds: Neo4jDatabaseService) = {
     entity.fields foreach {
       fieldObj => {
-        //@TODO More generic approach
         fieldObj match {
           case field: IntField =>
             node.setProperty(field.name, field.is.asInstanceOf[java.lang.Integer])
           case field: StringField =>
             node.setProperty(field.name, field.is)
-          //case field: DoubleField =>
-          //  vertex.field(field.name, field.is)
           case field: UuidField =>
             node.setProperty(field.name, field.is.toString)
           case field: LongField =>
             node.setProperty(field.name, field.is.asInstanceOf[java.lang.Long])
-          //case field:LocaleField =>
-          //  vertex.field(field.name, field.is.toString)
           case field:DateTimeField =>
             node.setProperty(field.name, field.valueToDB.asInstanceOf[java.lang.Long])
           case field:BooleanField =>
@@ -338,7 +338,6 @@ abstract class Neo4jGraphRepository[EntityType <: Neo4jGraphEntity]
    */
   protected def setEntityFields(entity: EntityType, node: Node, depth: Int = 0)(implicit ds: Neo4jDatabaseService) = {
     entity.fields foreach {
-      //@TODO More generic approach
       fieldObj => {
         fieldObj match {
           case field: IntField =>
@@ -353,8 +352,6 @@ abstract class Neo4jGraphRepository[EntityType <: Neo4jGraphEntity]
             field.set(util.UUID.fromString(node.getProperty(field.name).asInstanceOf[String]))
           case field: LongField =>
             field.set(node.getProperty(field.name).toString.toLong)
-          //case field: LocaleField =>
-          //  field.set(new Locale(vertex.field(field.name)))
           case field: DateTimeField =>
             field.valueFromDB(node.getProperty(field.name))
           case field: BooleanField =>
